refactor & code reusage
